+++ src/components/BooleanDisplay/index.ts
@@ -1,0 +1,3 @@
+import BooleanDisplay from "./src/BooleanDisplay.vue";
+
+export default BooleanDisplay;