+++ src/components/EchartsPie/index.ts
@@ -1,0 +1,3 @@
+import EchartsPie from "./src/Pie.vue";
+
+export default EchartsPie;