+++ src/components/FlowerShadow/index.ts
@@ -1,0 +1,3 @@
+import FlowerShadow from "./src/FlowerShadow.vue";
+
+export default FlowerShadow;