+++ src/components/Heart/index.ts
@@ -1,0 +1,3 @@
+import Heart from "./src/Heart.vue";
+
+export default Heart;