+++ src/components/cchList/index.ts
@@ -1,4 +1,4 @@
-import List from "./src/index.vue";
+import List from "./src/CchList.vue";
 import ListItem from "./src/ListItem.vue";
 
 export { List, ListItem };