+++ src/components/cchList/src/constants.ts
@@ -1,2 +1,7 @@
 import type { InjectionKey } from "vue";
-export const ListKey: InjectionKey<any> = Symbol("ListKey");
+
+export interface ListProvide {
+  col: number;
+}
+
+export const ListKey: InjectionKey<ListProvide> = Symbol("ListKey");