+++ src/components/cchPage/index.ts
@@ -1,3 +1,0 @@
-import Table from "./src/page";
-
-export default Table;