+++ src/views/demo/cssdemo/index.vue
@@ -23,7 +23,7 @@ import GridDemo2 from "./src/GridLayoutDemo/GridDemo2.vue";
 import BoxShadowDemo from "./src/BoxShadowDemo/index.vue";
 import ShadowDemo2 from "./src/BoxShadowDemo/ShadowDemo2.vue";
 import ShadowDemo3 from "./src/BoxShadowDemo/ShadowDemo3.vue";
-import FlowerShadow from "@/components/FlowerShadow/index.vue";
+import FlowerShadow from "@/components/FlowerShadow";
 </script>
 
 <style scoped></style>